refactor: collect git paths into a revParse module
